requirements.txt
chromadb==0.5.11
pandas==2.2.3
python-dotenv==1.0.1
selenium==4.26.1
srt==3.5.3
